Feat: A2 컴포넌트 저장 모달 구현
