+++ src/components/doc/common/InspectionCompleteButton.jsx
@@ -14,7 +14,6 @@ const Button = styled.button `
 
 const InspectionCompleteButton = ({showCommModal }) => {
   return (
-      // <Button type='submit' onClick={e => onStorage(e, form, path)}>검사 완료</Button>
       <Button type='submit' onClick={showCommModal}>검사 완료</Button>
   );
 };
